feat: 댓글 작성 폼을 CommentCreateForm으로 분리, router에 메서드를 추가하여 이동 시 스크롤이 상단 고정되도록 구현
- router/index.js에 scrollBehavior 로직을 추가하여 라우팅 시 스크롤이 상단에 고정되도록 구현함. 이 과정에서 미사용 파라미터 때문에 ESLint 수정이 발생하여 disable 주석을 달았음.
- CommentCreateForm을 만들어 댓글 입력란과 그 로직을 분리함.
- CommentCreateForm으로 로직이 분리된 것을 ArticleDetail에 반영함.
